Admin: thêm xóa sửa (còn CTKM + ...)

+++ database/migrations/2018_05_11_041134_create_table_san_pham.php
@@ -17,7 +17,7 @@ class CreateTableSanPham extends Migration
             $table->engine = 'InnoDB';
             $table->increments('id');
             $table->string('tensanpham');
-            $table->float('gia');
+            $table->unsignedInteger('gia');
             $table->integer('soluong');
             $table->string('mota', 10000);
             $table->string('hinhanh'); //tên hình đại điện của sản phẩm .*
